+++ src/components/BookingWizard/Step1/CalendarDatePicker/CalendarDatePicker.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/team-sapphire-variables';
+@import '../../../../styles/team-sapphire-variables';
 
 .calendarDateNumber {
     height: 30px;
