+++ src/components/BookingWizard/Step3/index.jsx
@@ -1,12 +1,12 @@
 import React, { useContext } from "react";
 import { useNavigate, Link } from "react-router-dom";
-import { BookingWizardContext } from "./context";
-import { updateForm, setProgressBarStep } from "./context/actions";
+import { BookingWizardContext } from "../context";
+import { step3Schema } from "../context/schema";
+import { setProgressBarStep, updateForm } from "../context/actions";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { step4Schema } from "./schema";
 
-const Step4 = () => {
+const Step3 = () => {
   const navigate = useNavigate();
   const [state, dispatch] = useContext(BookingWizardContext);
 
@@ -19,16 +19,16 @@ const Step4 = () => {
     formState: { errors },
   } = useForm({
     initialValues,
-    resolver: yupResolver(step4Schema),
+    resolver: yupResolver(step3Schema),
   });
 
   const onSubmit = (formData) => {
     dispatch(updateForm(formData));
-    dispatch(setProgressBarStep(5));
-    navigate("/booking/step-5");
+    dispatch(setProgressBarStep(4));
+    navigate("/booking/step-4");
   };
 
-  const goBack = () => dispatch(setProgressBarStep(3));
+  const goBack = () => dispatch(setProgressBarStep(2));
 
   return (
     <div className="container pt-3">
@@ -36,7 +36,7 @@ const Step4 = () => {
         <div className="row my-3 align-items-end">
           <div className="col-12 col-lg-2 p-0 mb-3 mb-lg-0 order-2 order-lg-1">
             <Link
-              to="/booking/step-3"
+              to="/booking/step-2"
               onClick={goBack}
               className="link-dark text-decoration-none"
             >
@@ -54,4 +54,4 @@ const Step4 = () => {
   );
 };
 
-export default Step4;
+export default Step3;
