+++ src/components/BookingWizard/Step5/Waiver/Waiver.module.scss
@@ -6,4 +6,3 @@
 .clearSignatureButton:hover {
     cursor: pointer;
 }
-
