+++ src/components/BookingWizard/index.jsx
@@ -8,9 +8,8 @@ import Step3 from "./Step3";
 import Step4 from "./Step4";
 import Step5 from "./Step5";
 import Step6 from "./Step6";
-import Checkout from "./Checkout";
-import CartPreview from "./CartPreview";
-import ProgressBar from "./ProgressBar";
+import CartPreview from "./common/CartPreview";
+import ProgressBar from "./common/ProgressBar";
 
 const BookingWizard = () => {
   const [state, dispatch] = useReducer(wizardReducer, initialState);
@@ -29,7 +28,7 @@ const BookingWizard = () => {
                 <Route path="/step-3" element={<Step3 />} />
                 <Route path="/step-4" element={<Step4 />} />
                 <Route path="/step-5" element={<Step5 />} />
-                <Route path="/checkout" element={<Checkout />} />
+                <Route path="/checkout" element={<Step6 />} />
               </Routes>
             </div>
           </div>
